Split form into a display component
